Modified the test for the create_message_level function.
Instead of using the list of instrument symbols corresponding to source 684, the previous implementation of the test was using the get_source_symbols_dict fixture. Therefore, instead of providing a list of instruments, it was providing a dictionary with source-code:list-of-instrument-symbols key-value pairs. By running the create_message_level_regex function on this dictionary, instead of the expected list of instrument symbols, was resulting in the function iterating over the dictionary keys and therefore producing a regex with the source-code corresponding to the iteration of that time, instead of the instrument symbol as expected.