Modified the regex pattern in the create_message_regex and create_instrument_specific_regex functions.
In the original implementation, the regex pattern to account for the variability of the expiry of the futures contract was written as F:FESX\\[A-Z][0-9][0-9] (example for the FESX futures contract). The pattern has been changed to F:FESX\\[A-Z][0-9]{2} to make use of the {} quantifiers, that in this case informs that the capital letter (as identified by the bracket group [A-Z]) will be followed by 2 numbers ([0-9]{2}).