Modified the retrieve_source_name_pairs function.
In the new implementation, the function returns a list of (source, instrument_name) tuples instead of a list of "source, instrument_name" strings. The reasoning behind the change, is to streamline the process of generating and writing the data to the csv file. Using the list of tuples as an output, allows to directly pass the generated list to the csv writer such that each tuple is then seen by the writer as a logical row unit, with the first element representing the sourceId and the second element, the RTSsymbol.