Modified the test of the retrieve_source_name_pairs function.
The change is meant to reflect the change introduced in the retrieve_source_name_pairs function, that now returns a list of (source_id, instrument_symbol) tuples instead of a list of strings. The test function now checks that the list generated by the function matches a list of tuples, instead of the previously implemented list of strings.