The discover_reference_data_files function has been modified into the search_files function.
The new implementation of the function generalise the scope of the function by introducing the possibility for the user to pass a search pattern that is then used to search the files. The docstring of the function has been modified to document the new functionalities introduced and an Examples section has been included to show some typical usages of the search pattern.