Extended the retrieve source_symbol_pairs function to now trigger a system exit with error code 1 in case of wrong file type.
During the tests of the command line interface, an issue emerged when, instead of passing to the program a bz2 formatted file, a txt formatted COREREF file was erroneously passed. Rightfully, the program raised an OSError, informing of an invalid data stream. However, from the exception alone, it was not clear what was really causing the issue. Since the issue occurred when testing a specific search pattern (**/COREREF_367*) that was targeting a file from a specific source, an inspection of the directory containing the file showed that inside the folder were two files: a bz2 file and a txt file that was previously created by extracting the content of the bz2 file in the folder. Since the retrieve_source_symbol_pairs function assumes that the passed file has a bz2 extension, when it encountered the txt formatted file, it did not recognize the file opened by the context handler as a proper stream and therefore raised the OSError exception.

To make the exception more informative, the body of the function has been modified with a try-except block. The new implementation of the function will try first to open the file and, if an OSError exception is raised, will exit the program and return an error message showing the file that the program was trying to open and that triggered the error and the file extension. This way, the program is interrupted and an actionable error message is provided.