Added additional test cases for the retrieve_source_symbol_pairs function.
The two new test cases are used to cover the newly introduced feature that, in case the retrieve_source-symbol_pairs receives as an input a COREREF file that is not bz2 compressed, interrupts the program with an exit code 1 and returns an error message that describes the file that triggered the interruption and the reason behind the failure. The two new tests case make sure that the system exit mechanism works as intended and that the message that accompanies the error corresponds to the expected message.