Added functions to create instrument-type-specific regex patterns.
Previously to this change, we only had the create_specific_instrument_regex function that, however, was producing regular expressions that could match only the symbol structure of simple future contracts. With the introduction of these news functions, we are now able to match the symbol structure of equities, futures and options, thus expanding the capabilities of the program of creating configuration files that can support multiple type of securities.