Added three new functions to create instrument-type specific regexes. Modified the create_specific_instrument_regex function. Adjusted the docstrings of the regex creating functions for futures and options.
The three new functions adds the functionality of creating regular expressions that match the symbol structure of fixed income products, indices and forward contracts.

The docstrings of the futures and options regex creating functions have been adjusted as the two previous version of the docstrings mixed the content of the two (the futures function had parts of the option docstrings and vice-versa), possibly due to a mistake in copying the structure of the docstring and not editing out the information not relevant for the specific function's documentation.

The create_specific_instrument_regex function has now been modified to adapt to the type of input symbol that it receive, and to produce as an output the correct regular expression that matches the type of the input. In its previous implementation, the function was generating only a regular expression that matched the structure of a futures contract. In the new implementation, the function infers the type from the type index in the inpyt symbol, and calls the right regex generating function matching the right type.