Modified some of the regex generating functions.
The equity, fixed income and indices specific regex generating functions have been modified after discovering, during a test of the equity regex generating function, that the specification of the regex pattern allowed matching also symbols that were containing the original root symbol prefixed by the type and session indicators (in the specific test, the regex pattern for the E:ADA symbol, was matching both E:ADA and E:ADAPTS, given that the first is a subset of the other). The respective functions have been now modified by adding the \b token, such that the passed root symbol is isolated and only exact matches are produced.