fix: Implement sorting of source-instrument pairs
Before the fix, the generated file did not have the source-instrument
pairs sorted by the source id as in the configuration file that is
generated when the Watchlist API is queried.

This commit fixes this behaviour by making sure that the discovered
source-instrument pairs are sorted before they are written to the
csv configuration file. This ensures that the configuration file that
is written at the end of the program has now all the source-instrument
pairs sorted in ascending order by the source id.